Add memory profiling support (requires a JDK patch)
The JDK patch adds support for a callback called everry X bytes of allocated memory.

There are some trivial changes to add configuration for memory profiling sample
size and to find/invoke the function to set the callback.

The bulk of the patch propagates the number of memory samples from the callback
to the circual buffer, to the log writer, to the file reader. The number of
samples is needed because when memory sample size is set (for example) to 65536
bytes, a single 3MB allocation counts for 48 samples.